refactor: user refactor commented by mentor

+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/_core/security/CustomUserDetails.java
@@ -22,7 +22,7 @@ public class CustomUserDetails implements UserDetails {
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return Collections.singleton(new SimpleGrantedAuthority(user.getDtype()));
     }
-    public Pair<String, Integer> getInfo() {
+    public Pair<String, Long> getInfo() {
         return Pair.of(user.getDtype(), user.getId());
     }
